requirements.txt
pandas
numpy
requests
tqdm
python-dotenv
plotly
streamlit
google-generativeai
duckduckgo-search
newspaper3k
lxml_html_clean
beautifulsoup4
html5lib
lxml
